src/requirements.txt
fastapi==0.111.0
pydantic-settings==2.10.1
uvicorn==0.38.0
jinja2==3.1.5
websockets==16.0